:recycle: refactor: 댓글 수 N+1문제 해결 #93

+++ src/main/java/com/sm/project/converter/community/PostConverter.java
@@ -79,7 +79,7 @@ public class PostConverter {
                     .nickname(post.getMember().getNickname())
                     .createdAt(post.getCreatedAt())
                     .itemImgUrlList(imgs)
-                    .commentCount(commentCountMap.getOrDefault(post.getId(), 0L).intValue())  //댓글 수 조회 -> 기본값 = 0
+                    .commentCount(commentCountMap.getOrDefault(post.getId(), 0L))  //댓글 수 조회 -> 기본값 = 0
                     .build();
         }).collect(Collectors.toList());
         return postDTOS;
